requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0

# Redis
redis[hiredis]>=5.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector Database
chromadb>=0.4.0
faiss-cpu>=1.7.4
weaviate-client>=3.25.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# LLM Integration
openai>=1.3.0
anthropics>=0.7.0
langchain>=0.0.350
langchain-community>=0.0.10
tiktoken>=0.5.0

# Knowledge Graph
neo4j>=5.14.0
networkx>=3.2.0
spacy>=3.7.0
nltk>=3.8.0

# GraphRAG
graphrag>=0.1.0
microsoft-graphrag>=0.1.0

# Monitoring
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0

# Message Queue
celery[redis]>=5.3.0
kombu>=5.3.0
pika>=1.3.0